# console/service/api/swagger.yaml
...
swagger: '2.0'
info:
  title: autobase console
  description: API for autobase console
  version: 2.8.0
host: localhost:8080
schemes:
  - http
produces:
  - application/json
consumes:
  - application/json
basePath: "/api/v1"
paths:
...
  /clusters:
    post:
      summary: "Create new cluster"
      tags:
        - cluster
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/Request.ClusterCreate'
      responses:
        '200':
          description: OK
          schema:
            $ref: "#/definitions/Response.ClusterCreate"
        '400':
          description: Error
          schema:
            $ref: "#/definitions/Response.Error"
    get:
      summary: "Get info about clusters"
      tags:
        - cluster
      parameters:
        - name: offset
          in: query
          required: false
          type: integer
        - name: limit
          in: query
          required: false
          type: integer
        - name: project_id
          in: query
          required: true
          type: integer
        - name: name
          in: query
          required: false
          type: string
          description: "Filter by name"
        - name: status
          type: string
          in: query
          required: false
          description: "Filter by status"
        - name: location
          type: string
          in: query
          required: false
          description: "Filter by location"
        - name: environment
          type: string
          in: query
          required: false
          description: "Filter by environment"
        - name: server_count
          type: integer
          in: query
          required: false
          description: "Filter by server_count"
        - name: postgres_version
          type: integer
          in: query
          required: false
          description: "Filter by postgres_version"
        - name: created_at_from
          required: false
          type: string
          format: date-time
          in: query
          description: "Created at after this date"
        - name: created_at_to
          required: false
          type: string
          format: date-time
          in: query
          description: "Created at till this date"
        - name: sort_by
          in: query
          required: false
          type: string
          description: "Sort by fields. Example: sort_by=id,-name,created_at,updated_at\n
          Supported values:\n
           - id\n
           - name\n
           - created_at\n
           - updated_at\n
           - environment\n
           - project\n
           - status\n
           - location\n
           - server_count\n
           - postgres_version\n"
      responses:
        '200':
          description: OK
          schema:
            $ref: "#/definitions/Response.ClustersInfo"
        '400':
          description: Error
          schema:
            $ref: "#/definitions/Response.Error"
...
  /clusters/{id}:
    get:
      summary: "Get cluster info"
      tags:
        - cluster
      parameters:
        - name: id
          in: path
          required: true
          type: integer
      responses:
        '200':
          description: OK
          schema:
            $ref: "#/definitions/ClusterInfo"
        '400':
          description: Error
          schema:
            $ref: "#/definitions/Response.Error"

    delete:
      summary: "Delete cluster (from the console database)"
      tags:
        - cluster
      parameters:
        - name: id
          in: path
          required: true
          type: integer
      responses:
        '204':
          description: OK
        '400':
          description: Error
          schema:
            $ref: "#/definitions/Response.Error"
...
  /clusters/{id}/operations:
    post:
      summary: "Launch a confirmed guarded cluster operation"
      tags:
        - cluster
      parameters:
        - name: id
          in: path
          required: true
          type: integer
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/Request.OperationStart'
      responses:
        '202':
          description: Accepted
          schema:
            $ref: '#/definitions/Response.OperationStart'
        '400':
          description: Error
          schema:
            $ref: '#/definitions/Response.Error'
...
  /operations:
    get:
      summary: "Get operations list for current project"
      tags:
        - operation
      parameters:
        - name: project_id
          in: query
          required: true
          type: integer
          description: "Required parameter for filter"
        - name: start_date
          required: true
          type: string
          format: date-time
          in: query
          description: "Operations started after this date"
        - name: end_date
          required: true
          type: string
          format: date-time
          in: query
          description: "Operations started till this date"
        - name: cluster_name
          in: query
          required: false
          type: string
          description: "Filter by cluster_name"
        - name: type
          in: query
          required: false
          type: string
          description: "Filter by type"
        - name: status
          in: query
          required: false
          type: string
          description: "Filter by status"
        - name: environment
          in: query
          required: false
          type: string
          description: "Filter by environment"
        - name: sort_by
          in: query
          required: false
          type: string
          description: "Sort by fields. Example: sort_by=cluster_name,type,status,id,started,finished\n
          Supported values:\n
           - id\n
           - cluster_name\n
           - type\n
           - status\n
           - started\n
           - finished\n
           - cluster\n
           - environment\n"
        - name: limit
          in: query
          required: false
          type: integer
        - name: offset
          in: query
          required: false
          type: integer
      responses:
        '200':
          description: OK
          schema:
            $ref: "#/definitions/Response.OperationsList"
        '400':
          description: Error
          schema:
            $ref: "#/definitions/Response.Error"
...
  /operations/{id}:
    get:
      summary: "Get operation audit record"
      tags:
        - operation
      parameters:
        - name: id
          in: path
          required: true
          type: integer
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Response.OperationDetail'
        '400':
          description: Error
          schema:
            $ref: '#/definitions/Response.Error'
definitions:
...
  Request.OperationStart:
    type: object
    required: [preflight_id, confirmation]
    properties:
      preflight_id:
        type: integer
      confirmation:
        type: string
...
  Response.OperationStart:
    type: object
    properties:
      operation_id:
        type: integer
      status:
        type: string
  Response.OperationDetail:
    type: object
    properties:
      id:
        type: integer
      cluster_id:
        type: integer
      type:
        type: string
      status:
        type: string
      actor:
        type: string
      sanitized_params:
        type: object
        additionalProperties: true
      preflight_snapshot:
        type: object
        additionalProperties: true
      plan:
        type: array
        items:
          type: string
      affected_nodes:
        type: array
        items:
          type: string
      final_verification:
        type: object
        additionalProperties: true
      safe_next_action:
        type: string
        x-nullable: true
      started:
        type: string
        format: date-time
      finished:
        type: string
        format: date-time
        x-nullable: true
...
  Response.Error:
    title: Error object
    type: object
    properties:
      code:
        type: integer
      title:
        type: string
      description:
        type: string
  Meta.Pagination:
    title: Pagination info for list requests
    type: object
    properties:
      offset:
        type: integer
        x-nullable: true
      limit:
        type: integer
        x-nullable: true
      count:
        type: integer
        x-nullable: true
...
  Request.ClusterCreate:
    type: object
    description: "Request struct for cluster creation"
    properties:
      name:
        type: string
        example: "drm-prod-pgcluster"
      description:
        type: string
        description: "Info about cluster"
      auth_info:
        type: object
        description: "Info for deployment system authorization"
        properties:
          secret_id:
            type: integer
            example: 1
      project_id:
        type: integer
        description: "Project for new cluster"
      environment_id:
        type: integer
        description: "Project environment"
      envs:
        type: array
        items:
          type: string
      extra_vars:
        type: object
        description: "Ansible extra vars (arbitrary JSON object)"
        additionalProperties: true
      existing_cluster:
        type: boolean
        default: false
        description: "Import an already-running cluster; skip Ansible-based deployment"
      query_analytics_enabled:
        type: boolean
        default: true
        description: "Enable query analytics for new PostgreSQL 14-18 clusters"
...
  Response.ClusterCreate:
    type: object
    description: "Response struct for cluster creation"
    properties:
      cluster_id:
        type: integer
        description: "unique code for cluster"
      operation_id:
        type: integer
        description: "operation id"
...
  Response.ClustersInfo:
    type: object
    properties:
      data:
        type: array
        items:
          $ref: '#/definitions/ClusterInfo'
      meta:
        $ref: '#/definitions/Meta.Pagination'
...
  ClusterInfo:
    type: object
    description: "Cluster info"
    properties:
      id:
        type: integer
      name:
        type: string
        example: "drm-prod-pgcluster"
      description:
        type: string
      status:
        type: string
        example: "healthy"
      creation_time:
        type: string
        format: date-time
        example: "16.10.2023T11:20:00Z"
      environment:
        type: string
        example: "production"
      servers:
        type: array
        items:
          $ref: "#/definitions/ClusterInfo.Instance"
      postgres_version:
        type: integer
        format: int32
        example: 15
      cluster_location:
        type: string
        description: "Code of location"
        example: "eu-north-1"
      project_name:
        type: string
        description: "Project for cluster"
      connection_info:
        type: object
      extra_vars:
        type: string
        description: "Ansible variables"
      inventory:
        type: string
        description: "Ansible inventory"
...
  ClusterInfo.Instance:
    type: object
    description: "Instance info for current cluster"
    properties:
      id:
        type: integer
      name:
        type: string
        example: "pgnode1"
      ip:
        type: string
        example: "10.128.64.141"
      status:
        type: string
      role:
        type: string
        example: "leader"
      timeline:
        type: integer
        format: int64
        example: 1
        x-nullable: true
      lag:
        type: integer
        format: int64
        example: 0
        x-nullable: true
      tags:
        type: object
      pending_restart:
        type: boolean
        example: false
        x-nullable: true
...
  Response.Operation:
    type: object
    properties:
      id:
        type: integer
        example: 1
      cluster_name:
        type: string
        example: "drm-prod-cluster"
      started:
        type: string
        format: date-time
        example: "16.10.2023T11:20:00Z"
      finished:
        type: string
        format: date-time
        example: "16.10.2023T11:20:00Z"
        x-nullable: true
      type:
        type: string
        example: "deploy"
      status:
        type: string
        example: "success"
      environment:
        type: string
        example: "production"
...
  Response.OperationsList:
    type: object
    properties:
      data:
        type: array
        items:
          $ref: '#/definitions/Response.Operation'
      meta:
        type: object
        $ref: '#/definitions/Meta.Pagination'
...